Add a MaterialResolver that takes a legacy material name and data and creates an ItemStack, performing any necessary conversions for 1.13 material flattening. WIP

+++ rpcapi/src/main/java/com/enjin/rpc/mappings/mappings/shop/Item.java
@@ -20,7 +20,7 @@ public class Item {
     private Map<Integer, Variable> variables;
     @Getter
     @SerializedName(value = "icon_damage")
-    private Integer iconDamage;
+    private Byte iconDamage;
     @Getter
     @SerializedName(value = "icon_item")
     private String iconItem;
